fix(runner): open the first available task by default
